+++ source/extensions/isaacsim.robot.wheeled_robots.nodes/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.1.2] - 2026-05-11
+### Fixed
+- Guard navigation nodes against invalid quaternion inputs when computing yaw from OmniGraph orientations.
+
 ## [0.1.1] - 2026-04-13
 ### Removed
 - Remove the `omni.isaac.ml_archive` dependency